refactor: conflict 해결
