+++ src/main/java/com/example/backend/domain/comment/controller/CommentCommandController.java
@@ -1,9 +1,10 @@
 package com.example.backend.domain.comment.controller;
 
+import com.example.backend.domain.comment.dto.CommentCreateRequest;
 import com.example.backend.domain.comment.dto.CommentResponse;
 import com.example.backend.domain.comment.dto.CommentUpdateRequest;
-import com.example.backend.domain.comment.service.CommentService;
 import com.example.backend.domain.auth.security.CustomUserDetails;
+import com.example.backend.domain.comment.service.command.CommentCommandService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import lombok.RequiredArgsConstructor;
@@ -13,10 +14,21 @@ import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/api/v1/comments")
-public class CommentController {
+@RequestMapping("/api/v1/posts/{postId}/comments")
+public class CommentCommandController {
 
-    private final CommentService commentService;
+    private final CommentCommandService commentCommandService;
+
+    @Operation(summary = "댓글 생성", description = "특정 게시글에 댓글을 생성합니다.")
+    @ApiResponse(responseCode = "200", description = "댓글 생성 성공")
+    @PostMapping
+    public ResponseEntity<CommentResponse> createComment(
+            @AuthenticationPrincipal CustomUserDetails userDetails,
+            @PathVariable Long postId,
+            @RequestBody CommentCreateRequest commentCreateRequest){
+        CommentResponse comment = commentCommandService.createComment(userDetails.getUserId(), postId, commentCreateRequest);
+        return ResponseEntity.ok(comment);
+    }
 
     @Operation(summary = "댓글 수정", description = "특정 댓글을 수정합니다.")
     @ApiResponse(responseCode = "200", description = "댓글 수정 성공")
@@ -25,7 +37,7 @@ public class CommentController {
             @PathVariable Long commentId,
             @RequestBody CommentUpdateRequest commentUpdateRequest,
             @AuthenticationPrincipal CustomUserDetails userDetails){
-        CommentResponse response = commentService.updateComment(commentId, userDetails.getUserId(), commentUpdateRequest);
+        CommentResponse response = commentCommandService.updateComment(commentId, userDetails.getUserId(), commentUpdateRequest);
         return ResponseEntity.ok(response);
     }
 
@@ -35,7 +47,7 @@ public class CommentController {
     public ResponseEntity<Void> deleteComment(
             @PathVariable Long commentId,
             @AuthenticationPrincipal CustomUserDetails userDetails){
-        commentService.deleteComment(commentId, userDetails.getUserId());
+        commentCommandService.deleteComment(commentId, userDetails.getUserId());
         return ResponseEntity.noContent().build();
     }
 }
